Completed update.

+++ backend/app/Http/Controllers/PostController.php
@@ -13,7 +13,7 @@ class PostController extends Controller
 
         return response()->json([
             'status' => 'success',
-            'posts'  => $posts
+            'posts'  => $posts,
         ]);
     }
 
@@ -29,11 +29,11 @@ class PostController extends Controller
 
         return response()->json([
             'status'    => 'success',
-            'message'   => 'Post created successfully!'
+            'message'   => 'Post created successfully!',
         ]);
     }
 
-    public function delete(int $post_id){
+    public function show(int $post_id){
 
         $post = Post::find($post_id);
 
@@ -44,11 +44,53 @@ class PostController extends Controller
             ], 404);
         }
 
+        return response()->json([
+            'status' => 'success',
+            'post'   => $post,
+        ]);
+    }
+
+    public function update(Request $request, int $post_id){
+
+        $request->validate([
+            'title'         => ['required', 'string', 'max:255'],
+            'description'   => ['required', 'string', 'max:65500'],
+            'category_id'   => ['required', 'numeric', 'integer', 'exists:categories,id'],
+        ]);
+
+        $post = Post::find($post_id);
+
+        if(!$post){
+            return response()->json([
+                'status' => 'failed',
+                'message' => 'Post not found!'
+            ], 404);
+        }
+
+        $post->update($request->only('category_id', 'title', 'description'));
+
+        return response()->json([
+            'status'    => 'success',
+            'message'   => 'Post updated successfully!',
+        ]);
+    }
+
+    public function delete(int $post_id){
+
+        $post = Post::find($post_id);
+
+        if(!$post){
+            return response()->json([
+                'status' => 'failed',
+                'message' => 'Post not found!',
+            ], 404);
+        }
+
         $post->delete();
 
         return response()->json([
             'status' => 'success',
-            'message' => 'Post deleted successfully!'
+            'message' => 'Post deleted successfully!',
         ]);
     }
 }
